Restructuring and Clean up
* Reorganized code into folders
* Added a separate icon for menu bar
* Removed separate view for popup and re-used the one from the main app
* Fixed a couple of typos in Credits.rtf
* Renamed Info.plist files as per target and moved those to the project root
* Removed unused code files
