--- WebShell/Base.lproj/WebShellPopupViewController.xib
@@ -1,70 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<document type="com.apple.InterfaceBuilder3.Cocoa.XIB" version="3.0" toolsVersion="10116" systemVersion="15F24b" targetRuntime="MacOSX.Cocoa" propertyAccessControl="none" useAutolayout="YES">
-    <dependencies>
-        <plugIn identifier="com.apple.InterfaceBuilder.CocoaPlugin" version="10116"/>
-        <plugIn identifier="com.apple.WebKitIBPlugin" version="10116"/>
-        <capability name="stacking Non-gravity area distributions on NSStackView" minToolsVersion="7.0" minSystemVersion="10.11"/>
-    </dependencies>
-    <objects>
-        <customObject id="-2" userLabel="File's Owner" customClass="PopupViewController" customModule="WebShell" customModuleProvider="target">
-            <connections>
-                <outlet property="mainWebView" destination="wZJ-HS-esW" id="FOh-zp-Iuh"/>
-                <outlet property="view" destination="Hz6-mo-xeY" id="0bl-1N-x8E"/>
-            </connections>
-        </customObject>
-        <customObject id="-1" userLabel="First Responder" customClass="FirstResponder"/>
-        <customObject id="-3" userLabel="Application" customClass="NSObject"/>
-        <customView id="Hz6-mo-xeY">
-            <rect key="frame" x="0.0" y="0.0" width="556" height="343"/>
-            <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMinY="YES"/>
-            <subviews>
-                <stackView distribution="fill" orientation="vertical" alignment="leading" horizontalStackHuggingPriority="249.99998474121094" verticalStackHuggingPriority="249.99998474121094" detachesHiddenViews="YES" translatesAutoresizingMaskIntoConstraints="NO" id="Ukq-X3-rkk">
-                    <rect key="frame" x="0.0" y="0.0" width="556" height="343"/>
-                    <subviews>
-                        <stackView distribution="fill" orientation="vertical" alignment="leading" horizontalStackHuggingPriority="249.99998474121094" verticalStackHuggingPriority="249.99998474121094" detachesHiddenViews="YES" translatesAutoresizingMaskIntoConstraints="NO" id="ZqJ-Sc-waI">
-                            <rect key="frame" x="0.0" y="0.0" width="556" height="343"/>
-                            <subviews>
-                                <stackView distribution="fill" orientation="vertical" alignment="leading" horizontalStackHuggingPriority="249.99998474121094" verticalStackHuggingPriority="249.99998474121094" detachesHiddenViews="YES" translatesAutoresizingMaskIntoConstraints="NO" id="xgP-A5-GIb">
-                                    <rect key="frame" x="0.0" y="0.0" width="556" height="343"/>
-                                    <subviews>
-                                        <webView translatesAutoresizingMaskIntoConstraints="NO" id="wZJ-HS-esW">
-                                            <rect key="frame" x="0.0" y="0.0" width="556" height="343"/>
-                                            <webPreferences key="preferences" defaultFontSize="16" defaultFixedFontSize="13" minimumFontSize="0">
-                                                <nil key="identifier"/>
-                                            </webPreferences>
-                                        </webView>
-                                    </subviews>
-                                    <visibilityPriorities>
-                                        <integer value="1000"/>
-                                    </visibilityPriorities>
-                                    <customSpacing>
-                                        <real value="3.4028234663852886e+38"/>
-                                    </customSpacing>
-                                </stackView>
-                            </subviews>
-                            <visibilityPriorities>
-                                <integer value="1000"/>
-                            </visibilityPriorities>
-                            <customSpacing>
-                                <real value="3.4028234663852886e+38"/>
-                            </customSpacing>
-                        </stackView>
-                    </subviews>
-                    <visibilityPriorities>
-                        <integer value="1000"/>
-                    </visibilityPriorities>
-                    <customSpacing>
-                        <real value="3.4028234663852886e+38"/>
-                    </customSpacing>
-                </stackView>
-            </subviews>
-            <constraints>
-                <constraint firstAttribute="trailing" secondItem="Ukq-X3-rkk" secondAttribute="trailing" id="3fU-9s-0aA"/>
-                <constraint firstItem="Ukq-X3-rkk" firstAttribute="top" secondItem="Hz6-mo-xeY" secondAttribute="top" id="LA1-cU-JPD"/>
-                <constraint firstItem="Ukq-X3-rkk" firstAttribute="leading" secondItem="Hz6-mo-xeY" secondAttribute="leading" id="d4r-e2-DiS"/>
-                <constraint firstAttribute="bottom" secondItem="Ukq-X3-rkk" secondAttribute="bottom" id="frm-Ah-e12"/>
-            </constraints>
-            <point key="canvasLocation" x="434" y="339.5"/>
-        </customView>
-    </objects>
-</document>
