+++ WebShell/Core/WSApplication.swift
