+++ WebShell/Core/WSBaseSettings.swift
