+++ WebShell/Core/WSPasswordManager.swift
