+++ WebShell/Core/WSStringExtension.swift
