+++ WebShell/Core/WSTrackpadGestures.swift
