+++ WebShell/Core/WSWebViewFunctions.swift
