+++ WebShell/Credits.rtf
@@ -1,12 +1,12 @@
-{\rtf1\ansi\ansicpg1252\cocoartf1404\cocoasubrtf460
+{\rtf1\ansi\ansicpg1252\cocoartf1561\cocoasubrtf200
 {\fonttbl\f0\fswiss\fcharset0 Helvetica;}
 {\colortbl;\red255\green255\blue255;}
-\paperw12240\paperh15840\margl1440\margr1440\vieww9000\viewh8400\viewkind0
+{\*\expandedcolortbl;;}
+\margl1440\margr1440\vieww9000\viewh8400\viewkind0
 \pard\tx566\tx1133\tx1700\tx2267\tx2834\tx3401\tx3968\tx4535\tx5102\tx5669\tx6236\tx6803\pardirnatural\partightenfactor0
 
-\f0\fs24 \cf0 WebShell is a powerfull application wich makes it easy to integrate your web-app to a fully working app.\
-\pard\tx566\tx1133\tx1700\tx2267\tx2834\tx3401\tx3968\tx4535\tx5102\tx5669\tx6236\tx6803\pardirnatural\partightenfactor0
-\cf0 WebShell is created with help of:\
+\f0\fs24 \cf0 WebShell is a powerful application which makes it easy to integrate your web-app to a fully working app.\
+WebShell is created with help of:\
  Randy  		- https://github.com/djyde \
  Wesley 		- https://github.com/wdg\
  DreamPiggy 	- https://github.com/lizhuoli1126\
